requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiogram_i18n==1.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
frozenlist==1.4.1
idna==3.8
magic-filter==1.0.12
multidict==6.1.0
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
redis==5.0.8
typing_extensions==4.12.2
yarl==1.11.1
